refactor: 로깅 관련 로직을 별도 클래스로 분리
- domain/log/log.handler.ts로 추가
